python-flask-todo-list/requirements.txt
anyio==3.6.2
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
emoji==2.1.0
Flask==2.2.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nile-api==0.2.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0